requirements.txt
# API Clients
git+https://github.com/Polymarket/py-clob-client.git
git+https://github.com/Polymarket/python-order-utils.git
websockets==12.0
httpx==0.27.0
requests==2.31.0

# Framework y utilidades
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1

# LLM y embeddings
sentence-transformers==2.5.1
openai==1.12.0

# Base de datos y cache
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
chromadb>=0.6.3

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Core dependencies
py-clob-client>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
websockets>=10.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-mock>=3.10.0

# Logging and monitoring
structlog>=21.5.0

# Database
asyncpg>=0.27.0
SQLAlchemy>=2.0.0
